+++ src/components/map-layer-group/o-map-layer-group.component.ts
@@ -1,8 +1,5 @@
-import { Component, Inject, forwardRef, ViewChildren } from '@angular/core';
-import {
-  OMapComponent,
-  OMapLayerComponent
-} from '../../components';
+import { Component, Inject, forwardRef, ViewChildren, ViewEncapsulation } from '@angular/core';
+import { OMapComponent, OMapLayerComponent } from '../../components';
 import { LayerConfiguration, LayerGroupConfiguration, OMapLayerGroupsWarehouse } from '../../core';
 
 @Component({
@@ -16,22 +13,27 @@ import { LayerConfiguration, LayerGroupConfiguration, OMapLayerGroupsWarehouse }
     'collapsed: group-collapsed',
     'refGroup: group-ref'
   ],
-  template: require('./o-map-layer-group.component.html'),
-  styles: [require('./o-map-layer-group.component.scss')]
+  templateUrl: './o-map-layer-group.component.html',
+  styleUrls: ['./o-map-layer-group.component.scss'],
+  encapsulation: ViewEncapsulation.None,
+  host: {
+    '[class.o-map-layer-group]': 'true'
+  }
 })
+
 export class OMapLayerGroupComponent {
   id: string;
   idParent: string;
   name: string;
-  description: string;
-  collapsed: boolean = false;
+  protected _description: string;
+  protected _collapsed: boolean = false;
   refGroup: LayerGroupConfiguration;
 
   public mapLayers: Array<LayerConfiguration> = new Array<LayerConfiguration>();
   public mLayerGroupsWarehouse: OMapLayerGroupsWarehouse = new OMapLayerGroupsWarehouse();
 
-  @ViewChildren(OMapLayerComponent) layerChildren: OMapLayerComponent[];
-  @ViewChildren(OMapLayerGroupComponent) layerGroupChildren: OMapLayerGroupComponent[];
+  @ViewChildren(forwardRef(() => OMapLayerComponent)) layerChildren: OMapLayerComponent[];
+  @ViewChildren(forwardRef(() => OMapLayerGroupComponent)) layerGroupChildren: OMapLayerGroupComponent[];
 
   constructor(
     @Inject(forwardRef(() => OMapComponent)) protected oMap: OMapComponent
@@ -64,7 +66,11 @@ export class OMapLayerGroupComponent {
     return childLayerGroupInWS || childLayerInWS;
   }
 
-  protected getLayerGroups() {
+  public getMapLayers() {
+    return this.mapLayers;
+  }
+
+  public getLayerGroups() {
     if (this.mLayerGroupsWarehouse.all &&
       this.mLayerGroupsWarehouse.all.length > 0) {
       return this.mLayerGroupsWarehouse.all;
@@ -80,8 +86,23 @@ export class OMapLayerGroupComponent {
     return center;
   }
 
-  private toggleCollapse() {
+  public toggleCollapse() {
     this.collapsed = !this.collapsed;
   }
 
+  get collapsed(): boolean {
+    return this._collapsed;
+  }
+
+  set collapsed(val: boolean) {
+    this._collapsed = val;
+  }
+
+  get description(): string {
+    return this._description;
+  }
+
+  set description(val: string) {
+    this._description = val;
+  }
 }