+++ src/components/map/o-map-w-events.class.ts
@@ -1,5 +1,5 @@
 import { EventEmitter } from '@angular/core';
-import * as L from 'leaflet';
+// import * as L from 'leaflet';
 import { OMapEvents } from './o-map-events.interface';
 import { OMapBase } from './o-map.class';
 
@@ -18,7 +18,7 @@ export class OMapWEvents extends OMapBase {
 		 * When configuration is triggered
 		 * this subscription maps Leaflet events to OMap EventEmitters
 		 */
-    this.mapConfiguration.subscribe(() => {
+    this.onMapConfigured().subscribe(() => {
       ['Click', 'Drag', 'Move', 'MoveEnd', 'ZoomLevelsChange'].forEach(eventName => {
         this.getMapService().map.on(
           eventName.toLowerCase(),