+++ src/components/toggle-icon-button/o-toggle-icon-button.component.ts
@@ -12,16 +12,16 @@ import { Util } from '../../utils';
     'clickable',
     'noInk: no-ink'
   ],
-  template: require('./o-toggle-icon-button.component.html'),
-  styles: [require('./o-toggle-icon-button.component.scss')]
+  templateUrl: './o-toggle-icon-button.component.html',
+  styleUrls: ['./o-toggle-icon-button.component.scss']
 })
 export class OToggleIconButtonComponent {
-  buttonActive: boolean = false;
-  iconDark: boolean;
-  iconName: string;
-  bgColor: string;
-  clickable: boolean;
-  noInk: boolean = false;
+  protected _buttonActive: boolean = false;
+  protected _iconDark: boolean;
+  protected _iconName: string;
+  protected bgColor: string;
+  protected clickable: boolean;
+  protected _noInk: boolean = false;
 
   ngOnInit() {
     this.clickable = Util.parseBoolean(this.clickable ? this.clickable.toString() : 'yes', true);
@@ -46,5 +46,37 @@ export class OToggleIconButtonComponent {
       return this.buttonActive = !this.buttonActive;
     }
     return false;
+  }
+
+  get buttonActive(): boolean {
+    return this._buttonActive;
+  }
+
+  set buttonActive(val: boolean) {
+    this._buttonActive = val;
+  }
+
+  get iconDark(): boolean {
+    return this._iconDark;
+  }
+
+  set iconDark(val: boolean) {
+    this._iconDark = val;
+  }
+
+  get iconName(): string {
+    return this._iconName;
+  }
+
+  set iconName(val: string) {
+    this._iconName = val;
+  }
+
+  get noInk(): boolean {
+    return this._noInk;
+  }
+
+  set noInk(val: boolean) {
+    this._noInk = val;
   }
 }