+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+##2.0.0-rc.2
+
+### Features
+* **ontimize-web-ng2-map**: '*OMapModule*' is now AoT compatible.
+
+### DEPENDENCY UPDATES ###
+* **Added**:  leaflet.heat@0.2.0
+
+
+### Bug Fixes
+* **o-map**: fixing errors when '*o-map*' is inside a tab ([#12](https://github.com/OntimizeWeb/ontimize-web-ng2-map/issues/12)) ([4ae8fbd](https://github.com/OntimizeWeb/ontimize-web-ng2-map/commit/4ae8fbd))
+* **o-navigator-item**: fixing search results alignment bug ([#13](https://github.com/OntimizeWeb/ontimize-web-ng2-map/issues/13)) ([370c0ab](https://github.com/OntimizeWeb/ontimize-web-ng2-map/commit/370c0ab))
+* **o-map**: fixing styles bugs ([#13](https://github.com/OntimizeWeb/ontimize-web-ng2-map/issues/13)) ([72c6f82](https://github.com/OntimizeWeb/ontimize-web-ng2-map/commit/72c6f82))
+* **o-map-layer-group**: fixing display problems ([#10](https://github.com/OntimizeWeb/ontimize-web-ng2-map/issues/10)) ([52a3cec](https://github.com/OntimizeWeb/ontimize-web-ng2-map/commit/52a3cec))
+
 ## 1.0.2
 ### Features
 * **o-map**: adding '*onToggleWSLayerSelected*', '*onToggleWSLayerVisibility*' and '*onToggleWSLayerInWS*' outputs.